Propagate type of node_data for comparison operations

+++ phylanx/util/slicing_helpers.hpp
@@ -29,7 +29,7 @@ namespace phylanx { namespace util { namespace slicing_helpers
 
     ///////////////////////////////////////////////////////////////////////////
     // generate a list of indices to extract from a given vector
-    PHYLANX_EXPORT std::vector<std::int64_t> create_list_slice(
+    PHYLANX_EXPORT std::vector<std::size_t> create_list_slice(
         std::int64_t start, std::int64_t stop, std::int64_t step);
 
     ///////////////////////////////////////////////////////////////////////////
